requirements.txt
google-generativeai
python-dotenv
tqdm
pandas
numpy
